pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bst"
version = "0.1.0"
description = "Bastion CLI for clinical-trial protocol version control"
authors = [{name = "bst contributors"}]
dependencies = [
  "click>=8.0",
  "PyYAML>=6.0",
  "jsonschema>=4.0",
  "GitPython>=3.1",
  "rich>=13.0"
]
requires-python = ">=3.8"
readme = "README.md"

[project.scripts]
bst = "bst.cli:cli"

[tool.setuptools.packages.find]
include = ["bst*"]
